requirements.txt
# Runtime dependencies (mirrors setup.py install_requires)
click>=8.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Development / testing
pytest>=7.0
